python-backend/requirements.txt
# FastAPI and Web Server
fastapi>=0.68.0
uvicorn>=0.15.0

# Database ORM and Migrations
sqlalchemy>=1.4.23
alembic>=1.7.1
psycopg2-binary>=2.9.1
pymongo>=4.6.1


# Data Validation and Settings Management
pydantic>=1.8.2
pydantic-settings>=2.0.0

# Environment Variables
python-dotenv>=0.19.0

# Authentication and Security
python-jose>=3.3.0
passlib>=1.7.4

# File Handling
python-multipart>=0.0.5

# Data Processing
pandas>=2.2.0

# Additional Dependencies for Production
bcrypt>=4.1.2  # For password hashing
email-validator>=2.1.0  # For email validation
python-dateutil>=2.8.2  # For date handling
openpyxl>=3.1.2  # For Excel file handling
xlrd>=2.0.1  # For older Excel file formats